refactor: remove utils-ol - functions are now in separate files
